requirements.txt
# Webview and Flask Server
pywebview
wtforms
Flask
flask-wtf

# For Macro Management
keyboard
mouse